remotely from Santos - Canal 2 - listening to orchestra of Reforma of Pizzeria Sollaviaggio
at inha at veriverse - at Dell XPS 8300 Intel(R) Core(TM) i7-2600 CPU @ 3.40GHz 16GB RAM 500GB + 500GB 7200rpm storage + 120 GB SSD OCZ Agility 3 with Windows 8 Pro - with Portuguese - Brazil- ~10Mpbs NET Virtua Internet at Alameda Princesa Izabel, 3000 - Curitiba - PR - Brazil

+++ appseed/ca/user/user_keyboard.cpp
@@ -360,7 +360,7 @@ namespace user
    void keyboard::translate_os_key_message(::ca::message::key * pkey)
    {
 
-      pkey->m_ekey = m_mapKey[pkey->m_nChar];
+      pkey->m_ekey = m_mapKey[(int32_t) pkey->m_nChar];
 
    }
